Move `diff.py` testing helpers to test tool module
This is now importable from Darker tests, too.
